hr_wallet/requirements.txt
Django==4.2.7
Pillow==10.1.0
whitenoise==6.6.0
djangorestframework==3.14.0
